Extract reusable MockEngine4 test double
Move the inline IBuildEngine4 test engine into a shared TestUtilities.MockEngine4
next to the existing MockEngine so other tests can reuse it.
