+++ src/Microsoft.Build.Tasks.Git.UnitTests/LocateRepositoryTests.cs
@@ -2,11 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the License.txt file in the project root for more information.
 
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.IO;
-using Microsoft.Build.Framework;
 using TestUtilities;
 using Xunit;
 
@@ -24,7 +20,7 @@ namespace Microsoft.Build.Tasks.Git.UnitTests
             using var temp = new TempRoot();
             var (repoDir, gitDir) = CreateMinimalRepository(temp);
 
-            var engine = new CachingMockEngine();
+            var engine = new MockEngine4();
             var task = new LocateRepository
             {
                 BuildEngine = engine,
@@ -60,7 +56,7 @@ namespace Microsoft.Build.Tasks.Git.UnitTests
                 // A CWD-dependent (pre-migration) implementation would resolve "." against this repository.
                 Directory.SetCurrentDirectory(repoDir.Path);
 
-                var engine = new CachingMockEngine();
+                var engine = new MockEngine4();
                 var task = new LocateRepository
                 {
                     BuildEngine = engine,
@@ -85,55 +81,6 @@ namespace Microsoft.Build.Tasks.Git.UnitTests
             gitDir.CreateFile("HEAD").WriteAllText("ref: refs/heads/main\n");
             gitDir.CreateFile("config").WriteAllText("");
             return (repoDir, gitDir);
-        }
-
-        /// <summary>
-        /// Minimal <see cref="IBuildEngine4"/> used only because <see cref="RepositoryTask"/> caches the located
-        /// repository via <c>BuildEngine4.Register/GetRegisteredTaskObject</c> (pre-existing behavior). The cache
-        /// misses on every lookup here, so each task instance performs a fresh lookup.
-        /// </summary>
-        private sealed class CachingMockEngine : IBuildEngine4
-        {
-            private readonly System.Text.StringBuilder _log = new();
-            private readonly Dictionary<object, object?> _registeredTaskObjects = new();
-
-            public string Log => _log.ToString();
-
-            public void LogErrorEvent(BuildErrorEventArgs e) => _log.AppendLine("ERROR: " + e.Message);
-            public void LogWarningEvent(BuildWarningEventArgs e) => _log.AppendLine("WARNING: " + e.Message);
-            public void LogMessageEvent(BuildMessageEventArgs e) => _log.AppendLine(e.Message);
-            public void LogCustomEvent(CustomBuildEventArgs e) => _log.AppendLine(e.Message);
-
-            public string ProjectFileOfTaskNode => "";
-            public int ColumnNumberOfTaskNode => 0;
-            public int LineNumberOfTaskNode => 0;
-            public bool ContinueOnError => true;
-            public bool IsRunningMultipleNodes => false;
-
-            public void RegisterTaskObject(object key, object? obj, RegisteredTaskObjectLifetime lifetime, bool allowEarlyCollection)
-                => _registeredTaskObjects[key] = obj;
-
-            public object? GetRegisteredTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
-                => _registeredTaskObjects.TryGetValue(key, out var value) ? value : null;
-
-            public object? UnregisterTaskObject(object key, RegisteredTaskObjectLifetime lifetime)
-            {
-                _registeredTaskObjects.TryGetValue(key, out var value);
-                _registeredTaskObjects.Remove(key);
-                return value;
-            }
-
-            public void Yield() { }
-            public void Reacquire() { }
-
-            public bool BuildProjectFile(string projectFileName, string[] targetNames, IDictionary globalProperties, IDictionary targetOutputs)
-                => throw new NotImplementedException();
-            public bool BuildProjectFile(string projectFileName, string[] targetNames, IDictionary globalProperties, IDictionary targetOutputs, string toolsVersion)
-                => throw new NotImplementedException();
-            public bool BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames, IDictionary[] globalProperties, IDictionary[] targetOutputsPerProject, string[] toolsVersion, bool useResultsCache, bool unloadProjectsOnCompletion)
-                => throw new NotImplementedException();
-            public BuildEngineResult BuildProjectFilesInParallel(string[] projectFileNames, string[] targetNames, IDictionary[] globalProperties, IList<string>[] removeGlobalProperties, string[] toolsVersion, bool returnTargetOutputs)
-                => throw new NotImplementedException();
         }
     }
 }
